requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
Brotli==1.1.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.2
cssselect2==0.7.0
defusedxml==0.8.0rc2
Django==5.0.2
django-cachalot==2.6.2
django-debug-toolbar==4.3.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.2.1
exceptiongroup==1.2.0
Faker==12.0.1
flower==2.0.1
fonttools==4.49.0
html5lib==1.1
humanize==4.9.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
install==1.3.5
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
mixer==7.2.2
oauthlib==3.2.2
packaging==23.2
pillow==10.2.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.8.0
PyJWT==2.8.0
pyphen==0.14.0
pytest==8.0.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.1
referencing==0.33.0
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.0
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.0
webencodings==0.5.1
zopfli==0.2.3
